Rename chatkeyboardview
