+++ src/components/ui/ChatKeyboardView.tsx
@@ -8,8 +8,8 @@ import {
   StyleSheet,
   View,
 } from 'react-native';
-import SendIcon from '../../components/ui/Icons/SendIcon/SendIcon';
-import SendIconDisabled from '../../components/ui/Icons/SendIcon/SendIconDisabled';
+import SendIcon from './Icons/SendIcon/SendIcon';
+import SendIconDisabled from './Icons/SendIcon/SendIconDisabled';
 import { vars } from '../../utils/theme';
 import CustomTextInput from './CustomTextInput';
